+++ test/core-e2e/src/route-variants.spec.js
@@ -195,7 +195,7 @@ describe("route variants", () => {
 
   describe("when restoring route variants", () => {
     it("should serve users under the /api/users path", async () => {
-      core.mocks.restoreRoutesVariants();
+      core.mocks.restoreRouteVariants();
       const users = await doFetch("/api/users?req=1");
       expect(users.status).toEqual(200);
       expect(users.body).toEqual([
@@ -285,7 +285,7 @@ describe("route variants", () => {
 
   describe("when restoring route variants again", () => {
     it("should serve user 2 under the /api/users/1 path", async () => {
-      core.mocks.restoreRoutesVariants();
+      core.mocks.restoreRouteVariants();
       const users = await doFetch("/api/users/1");
       expect(users.status).toEqual(200);
       expect(users.body).toEqual({ id: 2, name: "Jane Doe" });