+++ packages/core/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - feat(#334): Add new option `mock.collections.selected`. Deprecate `mocks.selected`
 - feat(#334): Deprecate `core.addRoutesHandler` method. Add alerts when used. `core.variantHandlers.register` method must be used instead.
 - feat(#334): Deprecate `routesHandlers` option. `variantHandlers.register` option must be used instead.
+- feat(#334): Deprecate `core.mock.restoreRoutesVariants` method. `core.mock.restoreRouteVariants` method must be used instead.
 
 ## [3.5.0] - 2022-07-05